2016 Day 11 [Perl]
Took a couple days to ignorantly come up with a depth-first search
without knowing it, then a few more days to learn about tree-search
algorithms (specifically A*) and Zobrist Hashing to correctly implement
this solution.

Was very pleasantly surprised to see my heuristic did so well on its
first revision.